requirements.txt
torch==2.4.1
torchvision==0.19.1
huggingface_hub
diffusers
ftfy
regex
tqdm
pytest
pyyaml
pandas==2.2.3
numpy>=2.1.0
einops==0.8.0
open_clip_torch==2.26.1
opencv-python>=4.11.0.0
transformers>=4.51
click
pyarrow
attrs
scikit-learn
accelerate
kornia
omegaconf==2.3.0